Support memmap in grh_reader
When reading grh, support interpreting a memmap.
