+++ satpy/tests/reader_tests/test_epsnative_reader.py
@@ -1,4 +1,27 @@
 # Copyright 2017-2022, European Organisation for the Exploitation of Meteorological Satellites (EUMETSAT)
+# Copyright (c) 2023 Satpy developers
+
+# This file is part of satpy.
+#
+# satpy is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# satpy is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# satpy.  If not, see <http://www.gnu.org/licenses/>.
+
+# This module is based on source code obtained from the
+# epct_plugin_gis package developed by B-Open Solutions srl for EUMETSAT under
+# contract EUM/C0/17/4600001943/0PN and released under Apache License
+# Version 2.0, January 2004, http://www.apache.org/licenses/.  The original
+# source including revision history and details on authorship can be found at
+# https://gitlab.eumetsat.int/open-source/data-tailor-plugins/epct_plugin_gis
+
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,11 +35,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# -*- coding: utf-8 -*-
+"""FIXME DOC."""
+
+import datetime
 import os
 
-from epct_plugin_gis import epsnative_reader
+import numpy as np
+import pytest
 
+from satpy.readers import epsnative_reader
 
 TEST_DATA_PATH = os.environ.get("EPCT_TEST_DATA_DIR", "")
 TEST_DATA = os.path.join(
@@ -51,6 +78,7 @@ SAMPLE_DATA = os.path.join(SAMPLE_PATH, "AVHRRL1.nat")
 
 
 def test_get_class_tuple():
+    """FIXME DOC."""
     class_string = "class_1_2_3"
     assert epsnative_reader.get_class_tuple(class_string) == ("class", 1, 2, 3)
 
@@ -59,6 +87,7 @@ def test_get_class_tuple():
 
 
 def test_csv_list_by_product():
+    """FIXME DOC."""
     csv_list = epsnative_reader.csv_list_by_product("AVHRRL1")
     exp_csv = ["giadr_1_3.csv", "giadr_2_2.csv", "mdr_2_5.csv", "mphr_0_2.csv"]
 
@@ -69,31 +98,42 @@ def test_csv_list_by_product():
 
 
 def test_assemble_descriptor():
+    """FIXME DOC."""
     descriptor = epsnative_reader.assemble_descriptor("AVHRRL1")
     exp_keys = {("giadr", 1, 3), ("giadr", 2, 2), ("mdr", 2, 5), ("mphr", 0, 2)}
 
     assert set(descriptor.keys()) == exp_keys
 
 
-def test_grh_reader():
-    grh = epsnative_reader.grh_reader(SAMPLE_DATA)
+@pytest.mark.parametrize("how", ["string", "file", "mmap"])
+def test_grh_reader(sample_file, how):
+    """FIXME DOC."""
+    if how == "file":
+        sample = open(sample_file, mode="rb")
+    elif how == "mmap":
+        sample = np.memmap(sample_file, mode="r", offset=0)
+    else:
+        sample = sample_file
 
+    grh = epsnative_reader.grh_reader(sample)
     assert len(grh) == 6
     assert grh[0] == "mphr"
     assert grh[1] == 0
     assert grh[2] == 2
     assert grh[3] == 3307
-    assert grh[4].isoformat() == "2018-01-22T13:28:03.130000"
-    assert grh[5].isoformat() == "2018-01-22T13:28:04.797000"
+    assert grh[4] == datetime.datetime(2019, 6, 5, 0, 23, 52, 653000)
+    assert grh[5] == datetime.datetime(2019, 6, 5, 2, 8, 56, 225000)
 
 
 def test_find_mphr_csv():
+    """FIXME DOC."""
     mphr_csv = epsnative_reader.find_mphr_csv()
 
     assert os.path.isfile(mphr_csv)
 
 
 def test_mphr_reader():
+    """FIXME DOC."""
     mphr_content = epsnative_reader.mphr_reader(SAMPLE_DATA)
 
     assert len(mphr_content) == 72
@@ -103,6 +143,7 @@ def test_mphr_reader():
 
 
 def test_first_class_occurrence():
+    """FIXME DOC."""
     class_name = "mphr"
     grh, offset = epsnative_reader.first_class_occurrence(SAMPLE_DATA, class_name)
     assert offset == 0
@@ -123,6 +164,7 @@ def test_first_class_occurrence():
 
 
 def test_read_ipr_sequence():
+    """FIXME DOC."""
     ipr_sequence = epsnative_reader.read_ipr_sequence(SAMPLE_DATA)
 
     assert len(ipr_sequence) == 11
@@ -131,6 +173,7 @@ def test_read_ipr_sequence():
 
 
 def test_read_grh_of_target_class():
+    """FIXME DOC."""
     target_offset = 0
     with open(SAMPLE_DATA, "rb") as eps_fileobj:
         target_grh = epsnative_reader.read_grh_of_target_class(eps_fileobj, target_offset)
@@ -149,6 +192,7 @@ def test_read_grh_of_target_class():
 
 
 def test_add_info_about_class():
+    """FIXME DOC."""
     mphr = epsnative_reader.mphr_reader(SAMPLE_DATA)
     current_class = {"class": "", "offset": 5077}
     with open(SAMPLE_DATA, "rb") as eps_fileobj:
@@ -160,6 +204,7 @@ def test_add_info_about_class():
 
 
 def test_reckon_dtype():
+    """FIXME DOC."""
     assert epsnative_reader.reckon_dtype("boolean") == ">i1"
     assert epsnative_reader.reckon_dtype("u-integer1") == ">u1"
     assert epsnative_reader.reckon_dtype("vinteger2") == "byte,>i2"
@@ -168,6 +213,7 @@ def test_reckon_dtype():
 
 
 def test_bands_to_records_reader():
+    """FIXME DOC."""
     records = epsnative_reader.bands_to_records_reader("AVHRRL1")
 
     assert len(records) == 7
@@ -177,6 +223,7 @@ def test_bands_to_records_reader():
 
 
 def test_add_record_info_to_band():
+    """FIXME DOC."""
     record_info = epsnative_reader.add_record_info_to_band("AVHRRL1")
 
     assert len(record_info) == 7
@@ -187,6 +234,7 @@ def test_add_record_info_to_band():
 
 
 def test_create_toc():
+    """FIXME DOC."""
     toc = epsnative_reader.create_toc(SAMPLE_DATA)
 
     assert len(toc) == 11
