+++ satpy/tests/reader_tests/test_iasisnd02.py
@@ -37,31 +37,11 @@
 """Test reading IASI L2 SND."""
 
 import datetime
-import pathlib
-import shutil
 
 import dask
 import numpy as np
-import pytest
-import requests
 
 from ..utils import CustomScheduler
-
-_url_sample_file = ("https://go.dwd-nextcloud.de/index.php/s/z87KfL72b9dM5xm/download/"
-                    "IASI_SND_02_M01_20190605002352Z_20190605020856Z_N_O_20190605011702Z.nat")
-
-
-@pytest.fixture(scope="module")
-def sample_file(tmp_path_factory):
-    """Obtain sample file."""
-    fn = pathlib.Path("/media/nas/x21308/IASI/IASI_SND_02_M01_20190605002352Z_20190605020856Z_N_O_20190605011702Z.nat")
-    if fn.exists():
-        return fn
-    fn = tmp_path_factory.mktemp("data") / "IASI_SND_02_M01_20190605002352Z_20190605020856Z_N_O_20190605011702Z.nat"
-    data = requests.get(_url_sample_file, stream=True)
-    with fn.open(mode="wb") as fp:
-        shutil.copyfileobj(data.raw, fp)
-    return fn
 
 
 def test_read_giadr(sample_file):
